Block size configuration is now scalar.
It was a vector for a historical reason. We first wanted to support
variable block size encoding, but so far we don't see any improvement
by using multiple block sizes. Now it is removed, and if we revisit
multiple block sizes again, this field should be used as a "primary"
block size and we should add another field for this purpose.