Optimize `auto_correlation` by parallelizing it in another way.
Previous auto-correlation is computed by storing a lagged window in a
SIMD vector, and multiply it with each sample.
The new way is split the signal into SIMD vectors, compute partial
accumulator for each lane (and each delay-time), and computes the sum
of them. This can improve the lane utilization rate since the previous
approach required an unnecessary multiplication due to 2^N-size of
the lagged window vector.

This commit also rewords "delay_sum" to "lagged_outer_prod_sum",
delay-sum in signal processing is usually referring a different
algorithm and this is actually sum of products between a signal
and a lagged version of it.