Implemented last frame truncation.
There is a design question whether this should be implenented as an
optional behavior, or not.  We think that this is okay to be
forced-default because one can still avoid this behavior by implementing
a custom `Source` type.

There was a significant degradation in the IRTF metrics in the
integration test. However, the degradation was confirmed not to be
attributed to this commit.

Closes #199.