`parser` is also moved into `component` sub-module.
This commit also moves tests in `component` to relevant sub-modules, and
adds support for noise generation with fixed seeds.

Additionally, there're a lot of lines added for instructing inlining. This
is because moving parsers to the other module misteriously changed the
inlining score and there was a considerable regression in benchmark
performance.